custom_nodes/comfyui-frame-interpolation/requirements-with-cupy.txt
torch
numpy
einops
opencv-contrib-python
kornia
scipy
Pillow
torchvision
tqdm
cupy-wheel